# backend/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the backend service",
        "title": "Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/v1/stack-definitions": {
            "get": {
                "description": "List all stack definitions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stack-definitions"
                ],
                "summary": "List stack definitions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.StackDefinition"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new stack definition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stack-definitions"
                ],
                "summary": "Create a stack definition",
                "parameters": [
                    {
                        "description": "Definition object",
                        "name": "definition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StackDefinition"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.StackDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/stack-definitions/import": {
            "post": {
                "description": "Import a stack definition from a portable JSON bundle, creating a new definition with fresh IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stack-definitions"
                ],
                "summary": "Import a stack definition",
                "parameters": [
                    {
                        "description": "Export bundle",
                        "name": "bundle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DefinitionExportBundle"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/stack-definitions/{id}": {
            "get": {
                "description": "Get a stack definition by ID, including its chart configurations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stack-definitions"
                ],
                "summary": "Get a stack definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Definition ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StackDefinition"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing stack definition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stack-definitions"
                ],
                "summary": "Update a stack definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Definition ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Definition object",
                        "name": "definition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StackDefinition"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StackDefinition"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a stack definition if no running instances link to it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stack-definitions"
                ],
                "summary": "Delete a stack definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Definition ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ChartConfigExportData": {
            "type": "object",
            "properties": {
                "chart_name": {
                    "type": "string"
                },
                "chart_path": {
                    "type": "string"
                },
                "chart_version": {
                    "type": "string"
                },
                "default_values": {
                    "type": "string"
                },
                "deploy_order": {
                    "type": "integer"
                },
                "repository_url": {
                    "type": "string"
                },
                "source_repo_url": {
                    "type": "string"
                }
            }
        },
...
        "handlers.DefinitionExportBundle": {
            "type": "object",
            "properties": {
                "charts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ChartConfigExportData"
                    }
                },
                "definition": {
                    "$ref": "#/definitions/handlers.DefinitionExportData"
                },
                "exported_at": {
                    "type": "string"
                },
                "schema_version": {
                    "type": "string"
                }
            }
        },
        "handlers.DefinitionExportData": {
            "type": "object",
            "properties": {
                "default_branch": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.StackDefinition": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "default_branch": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "source_template_id": {
                    "type": "string"
                },
                "source_template_version": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key (format: \"sk_\u003ckey\u003e\")",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT Bearer token (format: \"Bearer \u003ctoken\u003e\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}